Improved value precision for percents in positions.
git-svn-id: https://svn.ctdp.net/svn/rfdynhud/trunk@210 1c0d6ce8-a1e3-b74c-a49d-f790e2421cd5

+++ java/src/net/ctdp/rfdynhud/widgets/widget/__WPrivilegedAccess.java
@@ -26,7 +26,7 @@ public class __WPrivilegedAccess
         widget.onPropertyChanged( property, oldValue, newValue );
     }
     
-    public static final void onPositionChanged( RelativePositioning oldPositioning, float oldX, float oldY, RelativePositioning newPositioning, float newX, float newY, Widget widget )
+    public static final void onPositionChanged( RelativePositioning oldPositioning, int oldX, int oldY, RelativePositioning newPositioning, int newX, int newY, Widget widget )
     {
         widget.onPositionChanged( oldPositioning, oldX, oldY, newPositioning, newX, newY );
     }
